% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{cosine_basis}
\alias{cosine_basis}
\title{Evaluate cosine basis}
\usage{
cosine_basis(z, n_basis)
}
\arguments{
\item{z}{points at which to evaluate the basis}

\item{n_basis}{number of basis functions}
}
\value{
matrix of basis evaluations, each column corresponds to a
  basis function, each row corresponds to a z point
}
\description{
Evaluate cosine basis
}
