% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{.haar_phi}
\alias{.haar_phi}
\title{Evaluates Haar mother wavlet}
\usage{
.haar_phi(x)
}
\arguments{
\item{x}{float, value at which to evaluate the wavlet}
}
\value{
float; the Haar mother wavlet evaluated at x
}
\description{
Evaluates Haar mother wavlet
}
