% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{estimate_loss}
\alias{estimate_loss}
\title{Estimate loss for RFCDE}
\usage{
estimate_loss(forest, bandwidth = "auto", method = "oob",
  x_test = NULL, z_test = NULL)
}
\arguments{
\item{forest}{An RFCDE object}

\item{bandwidth}{(optional) A bandwidth or "auto" for automatic
bandwidth selection.}

\item{method}{(optional) A string: either "oob" for out-of-bag
weights or "validation" for a validation data set}

\item{x_test}{(optional) The test covariates if using
`method="test"`.}

\item{z_test}{(optional) The test responses if using
`method="test"`.}
}
\value{
The CDE loss for the kernel density estimate
}
\description{
Estimate loss for RFCDE
}
