% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{evaluate_basis}
\alias{evaluate_basis}
\title{Calculate basis functions for new observations}
\usage{
evaluate_basis(z, n_basis, system = "cosine")
}
\arguments{
\item{z}{matrix whose row correspond to observations and whose
columns correspond to dimensions. Assumes each column has been
scaled to [0, 1].}

\item{n_basis}{integer number of how many basis functions should be
calculated for each dimension. If not same length as the number
of columns of `z` will recycle.}

\item{system}{Basis system to be used. If not same length as the
number of columns of `z` will recycle.}
}
\value{
A matrix of dimension nrow(z) by prod(n_basis) with entries
  consisting of the first n_basis functions in the (tensor) basis
  evaluated at the points z
}
\description{
Calculate basis functions for new observations
}
