% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ForestRcpp}
\alias{ForestRcpp}
\title{Fit a random forest for CDE}
\arguments{
\item{x_train}{a matrix of training covariates.}

\item{z_basis}{a matrix of basis functions evaluations for training responses.}

\item{n_trees}{the number of trees in the forest.}

\item{mtry}{the number of candidate variables to try for each split.}

\item{node_size}{the minimum number of observations in a leaf node.}
}
\description{
Provides a wrapper to the C++ RFCDE implementation.
}
