% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{format.RFCDE}
\alias{format.RFCDE}
\title{Format method for RFCDE objects}
\usage{
\method{format}{RFCDE}(x, ...)
}
\arguments{
\item{x}{A RFCDE object.}

\item{...}{Other arguments to format}
}
\description{
Format method for RFCDE objects
}
