% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{haar_basis}
\alias{haar_basis}
\title{Evaluate haar basis}
\usage{
haar_basis(z, n_basis)
}
\arguments{
\item{z}{points at which to evaluate the basis}

\item{n_basis}{number of basis functions}
}
\value{
matrix of basis evaluations, each column corresponds to a
  basis function, each row corresponds to a z point
}
\description{
Evaluate haar basis
}
