% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{kde_estimate}
\alias{kde_estimate}
\title{Helper function for kernel density estimation.}
\usage{
kde_estimate(z_train, z_grid, weights, bandwidth = "plugin")
}
\arguments{
\item{z_train}{matrix of training responses.}

\item{z_grid}{matrix of grid points to evaluate densities.}

\item{weights}{vector of weights.}

\item{bandwidth}{(optional) Either "plugin" for bandwidth selection by
plug-in rule, "cv" for cross-validation, or a fixed bandwidth
value or matrix. Defaults to "plugin".}
}
\value{
A vector of the density estimated at z_grid
}
\description{
Helper function for kernel density estimation.
}
