% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.R
\name{kde_loss}
\alias{kde_loss}
\title{Calculate loss for KDE estimates}
\usage{
kde_loss(weights, z_train, z_test, bandwidth = "auto")
}
\arguments{
\item{weights}{A matrix of weights for the training points.}

\item{z_train}{A matrix of training responses}

\item{z_test}{A matrix of test responses}

\item{bandwidth}{(optional) A bandwidth or "auto" for automatic
bandwidth selection.}
}
\value{
The CDE loss for the kernel density estimate
}
\description{
Calculate loss for KDE estimates
}
