% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{oob_weights}
\alias{oob_weights}
\title{Calculate out-of-bag weights.}
\usage{
oob_weights(forest)
}
\arguments{
\item{forest}{A RFCDE object.}
}
\value{
A matrix of out-of-bag weights for the training data.
}
\description{
Calculate out-of-bag weights.
}
