% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE-package.R
\docType{package}
\name{RFCDE-package}
\alias{RFCDE-package}
\title{RFCDE: Random Forests for Conditional Density Estimation}
\description{
Fits random forest conditional density estimates using conditional
density loss for splits.
}
\author{
Taylor Pospisil
}
