% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{RFCDE}
\alias{RFCDE}
\title{Fits a conditional density estimate random forest to training data.}
\usage{
RFCDE(x_train, z_train, n_trees, mtry, node_size, n_basis,
  basis_system = "cosine")
}
\arguments{
\item{x_train}{a matrix of training covariates.}

\item{z_train}{a matrix of training responses.}

\item{n_trees}{the number of trees in the forest.}

\item{mtry}{the number of candidate variables to try for each
split.}

\item{node_size}{the minimum number of observations in a leaf node.}

\item{n_basis}{the number of basis functions used in split density
estimates.}

\item{basis_system}{the system of basis functions to use;
currently "cosine" and "Haar" are supported.}
}
\description{
Fits a conditional density estimate random forest to training data.
}
