% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{summary.RFCDE}
\alias{summary.RFCDE}
\title{Summary method for RFCDE objects}
\usage{
\method{summary}{RFCDE}(x, ...)
}
\arguments{
\item{x}{A RFCDE object.}

\item{...}{Other arguments to summary}
}
\description{
Summary method for RFCDE objects
}
