% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis_functions.R
\name{tensor_basis}
\alias{tensor_basis}
\title{Calculate tensor basis functions for new observations}
\usage{
tensor_basis(z, n_basis, systems = "Fourier")
}
\arguments{
\item{z}{matrix whose row correspond to observations and whose
columns correspond to dimensions. Assumes each column has been
scaled to [0, 1].}

\item{n_basis}{integer vector of how many basis functions should be
calculated for each dimension. If not same length as the number
of columns of `z` will recycle.}

\item{systems}{vector of strings indicating the basis system to be
used. Options are "cosine", "Fourier", "Haar", and "Daubechies".
If not same length as the number of columns of `z` will recycle.}
}
\value{
A matrix of dimension nrow(z) by prod(n_basis) with entries
  consisting of the first n_basis functions in the tensor basis
  evaluated at the points z
}
\description{
Calculate tensor basis functions for new observations
}
