% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{tune_constant_bandwidth}
\alias{tune_constant_bandwidth}
\title{Select a constant bandwidth to minimize CDE loss}
\usage{
tune_constant_bandwidth(forest, bandwidths)
}
\arguments{
\item{forest}{A RFCDE object}

\item{bandwidths}{A list of bandwidths}
}
\value{
The loss associated with each choice of bandwidth
}
\description{
Select a constant bandwidth to minimize CDE loss
}
