% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{variable_importance}
\alias{variable_importance}
\title{Calculate variable importance measures for RFCDE.}
\usage{
variable_importance(forest, type = c("count", "loss"))
}
\arguments{
\item{forest}{a RFCDE object}

\item{type}{the type of importance measure; options are "count" and
"loss".}
}
\description{
Calculate variable importance measures for RFCDE.
}
