% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{weights}
\alias{weights}
\title{Obtains weights from RFCDE object.}
\usage{
weights(forest, x_test)
}
\arguments{
\item{forest}{A RFCDE object.}

\item{x_test}{A vector of test covariates.}
}
\value{
A vector of weights counting the number of co-occurances in leaf nodes of the forest.
}
\description{
Provides weights for the training data reflecting co-occurance in
leaf nodes of the forest.
}
