% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{weights.RFCDE}
\alias{weights.RFCDE}
\title{Obtains weights from RFCDE object.}
\usage{
\method{weights}{RFCDE}(object, newdata, ...)
}
\arguments{
\item{object}{A RFCDE object.}

\item{newdata}{A vector of test covariates.}

\item{\dots}{Other arguments}
}
\value{
A vector of weights counting the number of co-occurances in
  leaf nodes of the forest.
}
\description{
Provides weights for the training data reflecting co-occurance in
leaf nodes of the forest.
}
