% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RFCDE.R
\name{kde_estimate}
\alias{kde_estimate}
\title{Helper function for kernel density estimation.}
\usage{
kde_estimate(z_train, z_grid, weights, bandwidth = NULL)
}
\arguments{
\item{z_train}{matrix of training covariates.}

\item{z_grid}{matrix of grid points to evaluate densities.}

\item{weights}{vector of weights.}

\item{bandwidth}{bandwidth value or matrix.}
}
\value{
A vector of the density estimated at z_grid
}
\description{
Helper function for kernel density estimation.
}
