% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDE.R
\name{select_bandwidth}
\alias{select_bandwidth}
\title{Helper function for selecting bandwidth for KDE}
\usage{
select_bandwidth(z, weights, method = "plugin")
}
\arguments{
\item{z}{A matrix of training responses.}

\item{weights}{A vector of weights for training points.}

\item{method}{(optional) Either "plugin" for bandwidth selection by
plug-in rule, "cv" for cross-validation, or a fixed bandwidth
value or matrix. Defaults to "plugin".}
}
\value{
A bandwidth for KDE
}
\description{
Helper function for selecting bandwidth for KDE
}
